refactor: inline validateRepositoryDomain into repo-create.ts
It's only used in one place (the createRepo function), so it doesn't
need its own lib file.
