+++ src/commands/repo-create.ts
@@ -3,7 +3,18 @@ import { getHost, getToken } from "../auth";
 import { checkIsDomainAvailable, createRepository } from "../clients/wroom";
 import { CommandError, createCommand, type CommandConfig } from "../lib/command";
 import { UnknownRequestError } from "../lib/request";
-import { validateRepositoryDomain } from "../lib/repositoryDomain";
+
+const DOMAIN_REGEX = /^[a-zA-Z0-9][-a-zA-Z0-9]{2,}[a-zA-Z0-9]$/;
+const MIN_LENGTH = 4;
+const MAX_LENGTH = 63;
+
+export function validateRepositoryDomain(name: string): void {
+	if (name.length < MIN_LENGTH || name.length > MAX_LENGTH || !DOMAIN_REGEX.test(name)) {
+		throw new CommandError(
+			`Invalid repository name "${name}". Must be ${MIN_LENGTH}–${MAX_LENGTH} characters, letters/numbers/hyphens only, and start and end with a letter or number.`,
+		);
+	}
+}
 
 const config = {
 	name: "prismic repo create",
